requirements.txt
annotated-types==0.7.0
anyio==4.9.0
appier==1.34.6
cachetools==5.5.2
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.2.2
distro==1.9.0
fastapi==0.116.1
google-api-core==2.25.1
google-api-python-client==2.177.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
googleapis-common-protos==1.70.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
idna==3.10
jiter==0.10.0
numpy==2.3.2
oauthlib==3.3.1
openai==1.98.0
pandas==2.3.1
proto-plus==1.26.1
protobuf==6.31.1
pyactiveresource==2.2.2
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.7
pydantic_core==2.33.2
PyJWT==2.10.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
requests==2.32.4
requests-oauthlib==2.0.0
rsa==4.9.1
shopify-api==0.7.1
ShopifyAPI==12.7.0
six==1.17.0
sniffio==1.3.1
starlette==0.47.2
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.35.0
